backend/requirements.txt
sqlalchemy==2.0.35
fastapi==0.115.0
uvicorn==0.30.6
psycopg2-binary==2.9.3
google-genai==1.5.0
pydantic==2.10.6
pandas==2.2.3
numpy==2.2.4
pytest==8.3.5
